pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.1.3"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"
SQLAlchemy = "^1.4.39"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.1"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
ipython = "^8.4.0"
flake8 = "^4.0.1"
requests-mock = "^1.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
